Created and updated a util py script to offload some functions from main scripts and for better life cycle
